admonition: prefer button instead of X
